Configuration Property Files are now placed outside of JAR files.
Changed the message consumption mode of TaskExecutor, from async to sync. This was to fix a bug with the load balancing strategy. When using async mode, the load is balanaced according to the # of workers at the JOB START. With this change, it will allow dynamic
load balancing as intended.