Implemented Simple API.
Added a withhold duration for results if a task finishes before 2secs, which causes thread interaction issues.

Decreased Heartbeat period to 5 Seconds and increased the heartbeat thread priority to MAX.

Fixed a bug which causes not to reset the active job, when the cluster fails

Fixed a minor bug which allows multiple threads attempting to start grid node / cluster. 

This caused one thread to fail with an exception. Fixed by Synchronizing all start methods of Grid class on the class.

Fixed a bug in FutureServerImpl, where concurrect access to Job Listeners cause 

ConcurrentModificationException. Fixed by using a sychronized list for holding data and synchronizing the notfication process (iterator).

Fixed a minor bug in UI which hides the UI when user clicks close and then cancels it.

Fixed a minor bug in UnboundedJobProcessor which redefines the MAX_CONSEC_FAILURES field which is also defined in the super class.

Introduced UnboundedGridJobPostProcessAdapter, which provides a default implement for processResult() method.